Change layout, add mouseenter and radio buttons
Changed the layout to static 512px by 512px. The grids
change the height and width depending on the number
of grids the user wants.

Also added the mouseeenter event listener to each grid.
The grid changes color when mouse enters the grid. The
colors have two options based on the radio button selected
by the user. The two buttons correspond to grayscale and
RGB mode.